Move `types/query-language` to `common/query-language`
And fix imports
