+++ extensions/ql-vscode/src/common/query-language.ts
