+++ extensions/ql-vscode/test/factories/variant-analysis/shared/variant-analysis-submission.ts
@@ -1,6 +1,6 @@
 import { faker } from "@faker-js/faker";
 import { VariantAnalysisSubmission } from "../../../../src/variant-analysis/shared/variant-analysis";
-import { QueryLanguage } from "../../../../src/types/query-language";
+import { QueryLanguage } from "../../../../src/common/query-language";
 
 export function createMockSubmission(): VariantAnalysisSubmission {
   return {
