+++ extensions/ql-vscode/test/unit-tests/variant-analysis.test.ts
@@ -8,7 +8,7 @@ import {
 } from "../../src/variant-analysis/shared/variant-analysis";
 import { createMockScannedRepo } from "../factories/variant-analysis/shared/scanned-repositories";
 import { createMockVariantAnalysis } from "../factories/variant-analysis/shared/variant-analysis";
-import { QueryLanguage } from "../../src/types/query-language";
+import { QueryLanguage } from "../../src/common/query-language";
 
 describe("parseVariantAnalysisQueryLanguage", () => {
   it("parses a valid language", () => {
