+++ extensions/ql-vscode/test/vscode-tests/no-workspace/helpers.test.ts
@@ -30,7 +30,7 @@ import {
   walkDirectory,
 } from "../../../src/helpers";
 import { reportStreamProgress } from "../../../src/commandRunner";
-import { QueryLanguage } from "../../../src/types/query-language";
+import { QueryLanguage } from "../../../src/common/query-languagee";
 
 describe("helpers", () => {
   describe("Invocation rate limiter", () => {
